requirements.txt
moderngl==5.9.0
pygame_ce==2.3.1
